# build/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"CreatedIdResponse": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"username": {
						"type": "string"
					},
					"email": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"username",
					"email"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_CreatedIdResponse_": {
				"properties": {
					"success": {
						"type": "boolean",
						"enum": [
							true
						],
						"nullable": false
					},
					"message": {
						"type": "string"
					},
					"data": {
						"$ref": "#/components/schemas/CreatedIdResponse"
					}
				},
				"required": [
					"success",
					"message",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiErrorResponse": {
				"properties": {
					"success": {
						"type": "boolean",
						"enum": [
							false
						],
						"nullable": false
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"success",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"InitStaffBody": {
				"properties": {
					"username": {
						"type": "string",
						"example": "admin"
					},
					"first_name": {
						"type": "string",
						"example": "Tai"
					},
					"last_name": {
						"type": "string",
						"example": "Tran"
					},
					"email": {
						"type": "string",
						"example": "admin@example.com"
					},
					"password": {
						"type": "string",
						"example": "Aa@123456"
					}
				},
				"required": [
					"username",
					"first_name",
					"last_name",
					"email",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"StaffResponse": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"username": {
						"type": "string"
					},
					"first_name": {
						"type": "string"
					},
					"last_name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"phone": {
						"type": "string",
						"nullable": true
					},
					"address": {
						"type": "string",
						"nullable": true
					},
					"avatar": {
						"type": "string",
						"nullable": true
					},
					"status": {
						"type": "string",
						"enum": [
							"active",
							"inactive"
						]
					},
					"created_at": {
						"type": "string",
						"format": "date-time"
					},
					"updated_at": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"username",
					"first_name",
					"last_name",
					"email",
					"phone",
					"address",
					"avatar",
					"status",
					"created_at",
					"updated_at"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_StaffResponse_": {
				"properties": {
					"success": {
						"type": "boolean",
						"enum": [
							true
						],
						"nullable": false
					},
					"message": {
						"type": "string"
					},
					"data": {
						"$ref": "#/components/schemas/StaffResponse"
					}
				},
				"required": [
					"success",
					"message",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiMessageResponse": {
				"properties": {
					"success": {
						"type": "boolean",
						"enum": [
							true
						],
						"nullable": false
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"success",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"StaffUpdateProfileBody": {
				"properties": {
					"first_name": {
						"type": "string",
						"example": "Tai"
					},
					"last_name": {
						"type": "string",
						"example": "Tran"
					},
					"email": {
						"type": "string",
						"example": "admin@example.com"
					},
					"phone": {
						"type": "string",
						"example": "0778946513"
					},
					"address": {
						"type": "string",
						"example": "51 Duong A, TPHCM"
					},
					"avatar": {
						"type": "string",
						"example": "https://example.com/avatar.png"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"PaginationMeta": {
				"properties": {
					"total": {
						"type": "number",
						"format": "double"
					},
					"page": {
						"type": "number",
						"format": "double"
					},
					"limit": {
						"type": "number",
						"format": "double"
					},
					"totalPages": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"total",
					"page",
					"limit",
					"totalPages"
				],
				"type": "object",
				"additionalProperties": false
			},
			"StaffListResponse": {
				"properties": {
					"staff": {
						"items": {
							"$ref": "#/components/schemas/StaffResponse"
						},
						"type": "array"
					},
					"pagination": {
						"$ref": "#/components/schemas/PaginationMeta"
					}
				},
				"required": [
					"staff",
					"pagination"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_StaffListResponse_": {
				"properties": {
					"success": {
						"type": "boolean",
						"enum": [
							true
						],
						"nullable": false
					},
					"message": {
						"type": "string"
					},
					"data": {
						"$ref": "#/components/schemas/StaffListResponse"
					}
				},
				"required": [
					"success",
					"message",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateStaffBody": {
				"properties": {
					"username": {
						"type": "string",
						"example": "staff01"
					},
					"first_name": {
						"type": "string",
						"example": "Tai"
					},
					"last_name": {
						"type": "string",
						"example": "Tran"
					},
					"email": {
						"type": "string",
						"example": "staff01@example.com"
					},
					"password": {
						"type": "string",
						"example": "Aa@123456"
					}
				},
				"required": [
					"username",
					"first_name",
					"last_name",
					"email",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"StaffChangePasswordBody": {
				"properties": {
					"oldPassword": {
						"type": "string",
						"example": "Aa@123456"
					},
					"newPassword": {
						"type": "string",
						"example": "Aa@1234567"
					}
				},
				"required": [
					"oldPassword",
					"newPassword"
				],
				"type": "object",
				"additionalProperties": false
			},
			"StaffResetPasswordBody": {
				"properties": {
					"newPassword": {
						"type": "string",
						"example": "Aa@123456"
					}
				},
				"required": [
					"newPassword"
				],
				"type": "object",
				"additionalProperties": false
			},
			"StaffUpdateStatusBody": {
				"properties": {
					"status": {
						"type": "string",
						"enum": [
							"active",
							"inactive"
						],
						"example": "inactive"
					}
				},
				"required": [
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"ApiResponse_CustomerResponse_": {
				"properties": {
					"success": {
						"type": "boolean",
						"enum": [
							true
						],
						"nullable": false
					},
					"message": {
						"type": "string"
					},
					"data": {
						"$ref": "#/components/schemas/CustomerResponse"
					}
				},
				"required": [
					"success",
					"message",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ClientUpdateProfileBody": {
				"properties": {
					"first_name": {
						"type": "string",
						"example": "Tran"
					},
					"last_name": {
						"type": "string",
						"example": "Tai"
					},
					"email": {
						"type": "string",
						"example": "newemail@example.com"
					},
					"phone": {
						"type": "string",
						"example": "0901234567"
					},
					"address": {
						"type": "string",
						"example": "123 Le Loi, Q1, TPHCM"
					},
					"avatar": {
						"type": "string",
						"example": "https://example.com/avatar.png"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"ClientChangePasswordBody": {
				"properties": {
					"current_password": {
						"type": "string",
						"example": "OldPass@123"
					},
					"new_password": {
						"type": "string",
						"example": "NewPass@456"
					}
				},
				"required": [
					"current_password",
					"new_password"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"UpdateCustomerStatusBody": {
				"properties": {
					"status": {
						"type": "string",
						"enum": [
							"active",
							"inactive"
						],
						"example": "inactive"
					}
				},
				"required": [
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AdminResetPasswordBody": {
				"properties": {
					"new_password": {
						"type": "string",
						"example": "NewPass@123"
					}
				},
				"required": [
					"new_password"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "BMD Training API",
		"version": "1.0.0",
		"description": "API documentation for BMD Training backend",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
		"/api/admin/staff/init-staff": {
			"post": {
				"operationId": "InitStaff",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/ApiResponse_CreatedIdResponse_"
										},
										{
											"$ref": "#/components/schemas/ApiErrorResponse"
										}
									]
								}
							}
						}
					}
				},
				"summary": "Initialize the first staff account",
				"tags": [
					"Admin - Staff"
				],
				"security": [],
				"parameters": [
					{
						"in": "header",
						"name": "x-init-secret",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/InitStaffBody"
							}
						}
					}
				}
			}
		},
		"/api/admin/staff/profile": {
			"get": {
				"operationId": "GetProfile",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/ApiResponse_StaffResponse_"
										},
										{
											"$ref": "#/components/schemas/ApiErrorResponse"
										}
									]
								}
							}
						}
					}
				},
				"summary": "Get authenticated staff profile",
				"tags": [
					"Admin - Staff"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": []
			},
			"put": {
				"operationId": "UpdateProfile",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/ApiMessageResponse"
										},
										{
											"$ref": "#/components/schemas/ApiErrorResponse"
										}
									]
								}
							}
						}
					}
				},
				"summary": "Update authenticated staff profile",
				"tags": [
					"Admin - Staff"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/StaffUpdateProfileBody"
							}
						}
					}
				}
			}
		},
		"/api/admin/staff": {
			"get": {
				"operationId": "GetAllStaffProfile",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/ApiResponse_StaffListResponse_"
										},
										{
											"$ref": "#/components/schemas/ApiErrorResponse"
										}
									]
								}
							}
						}
					}
				},
				"summary": "Get all staff profiles",
				"tags": [
					"Admin - Staff"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"default": 1,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"default": 20,
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateStaff",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/ApiResponse_CreatedIdResponse_"
										},
										{
											"$ref": "#/components/schemas/ApiErrorResponse"
										}
									]
								}
							}
						}
					}
				},
				"summary": "Create a new staff account",
				"tags": [
					"Admin - Staff"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateStaffBody"
							}
						}
					}
				}
			}
		},
		"/api/admin/staff/{staffId}": {
			"get": {
				"operationId": "GetStaffById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/ApiResponse_StaffResponse_"
										},
										{
											"$ref": "#/components/schemas/ApiErrorResponse"
										}
									]
								}
							}
						}
					}
				},
				"summary": "Get a staff member by ID",
				"tags": [
					"Admin - Staff"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "staffId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/api/admin/staff/profile/change-password": {
			"put": {
				"operationId": "ChangePassword",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/ApiMessageResponse"
										},
										{
											"$ref": "#/components/schemas/ApiErrorResponse"
										}
									]
								}
							}
						}
					}
				},
				"summary": "Change password for authenticated staff",
				"tags": [
					"Admin - Staff"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/StaffChangePasswordBody"
							}
						}
					}
				}
			}
		},
		"/api/admin/staff/reset-password": {
			"put": {
				"operationId": "ResetPassword",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/ApiMessageResponse"
										},
										{
											"$ref": "#/components/schemas/ApiErrorResponse"
										}
									]
								}
							}
						}
					}
				},
				"summary": "Reset password for a staff member by ID",
				"tags": [
					"Admin - Staff"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/StaffResetPasswordBody"
							}
						}
					}
				}
			}
		},
		"/api/admin/staff/status": {
			"put": {
				"operationId": "UpdateStatus",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/ApiMessageResponse"
										},
										{
											"$ref": "#/components/schemas/ApiErrorResponse"
										}
									]
								}
							}
						}
					}
				},
				"summary": "Update status of a staff member by ID",
				"tags": [
					"Admin - Staff"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/StaffUpdateStatusBody"
							}
						}
					}
				}
			}
		},
...
		"/api/client/profile": {
			"get": {
				"operationId": "GetProfile",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/ApiResponse_CustomerResponse_"
										},
										{
											"$ref": "#/components/schemas/ApiErrorResponse"
										}
									]
								}
							}
						}
					}
				},
				"summary": "Get my profile",
				"tags": [
					"Client - Profile"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": []
			},
			"put": {
				"operationId": "UpdateProfile",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/ApiMessageResponse"
										},
										{
											"$ref": "#/components/schemas/ApiErrorResponse"
										}
									]
								}
							}
						}
					}
				},
				"summary": "Update my profile",
				"tags": [
					"Client - Profile"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ClientUpdateProfileBody"
							}
						}
					}
				}
			}
		},
		"/api/client/change-password": {
			"put": {
				"operationId": "ChangePassword",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/ApiMessageResponse"
										},
										{
											"$ref": "#/components/schemas/ApiErrorResponse"
										}
									]
								}
							}
						}
					}
				},
				"summary": "Change my password",
				"tags": [
					"Client - Profile"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ClientChangePasswordBody"
							}
						}
					}
				}
			}
		},
...
		"/api/admin/customers/status": {
			"put": {
				"operationId": "UpdateStatus",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/ApiMessageResponse"
										},
										{
											"$ref": "#/components/schemas/ApiErrorResponse"
										}
									]
								}
							}
						}
					}
				},
				"summary": "Update customer status",
				"tags": [
					"Admin - Client"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateCustomerStatusBody"
							}
						}
					}
				}
			}
		},
...
		"/api/admin/customers/reset-password": {
			"put": {
				"operationId": "ResetPassword",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/ApiMessageResponse"
										},
										{
											"$ref": "#/components/schemas/ApiErrorResponse"
										}
									]
								}
							}
						}
					}
				},
				"summary": "Reset a customer's password (admin only)",
				"tags": [
					"Admin - Client"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AdminResetPasswordBody"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}